backend/requirements.txt
# Core Flask dependencies
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-socketio==5.3.4

# Database
sqlalchemy==2.0.23

# Image processing and OCR
pytesseract==0.3.10
opencv-python==4.9.0.80
Pillow==10.2.0
passporteye==2.2.2

# Data processing and ML
numpy==1.26.2
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.1

# Optional ML (conditionally imported)
tensorflow==2.13.0

# Visualization
matplotlib==3.7.1
seaborn==0.12.2

# Authentication and security
PyJWT==2.8.0
cryptography==41.0.8
bcrypt==4.0.1

# Caching
redis==5.0.1

# WebSocket
python-socketio==5.8.0

# Monitoring
prometheus-client==0.19.0

# HTTP and utilities
requests==2.31.0
python-dotenv==1.0.0

# Production server
gunicorn==21.2.0

# Async task processing
celery==5.3.4
kombu==5.3.4
amqp==5.2.0
celery[redis]==5.3.4

# API documentation
flask-restx==1.3.0
flask-swagger-ui==4.11.1

# Enhanced monitoring and logging
structlog==24.1.0
python-json-logger==2.0.7
elastic-apm==6.20.0

# Testing
pytest==7.4.4
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
faker==22.0.0

# MCP (Model Context Protocol) dependencies
# Note: These are conceptual dependencies for MCP functionality
# The actual MCP implementation uses built-in Python libraries
dataclasses  # Built-in Python 3.7+
pickle  # Built-in Python
pathlib  # Built-in Python 3.4+
uuid  # Built-in Python
collections  # Built-in Python
enum  # Built-in Python 3.4+
mimetypes  # Built-in Python
hashlib  # Built-in Python